refactored js
