+++ fitnesse/resources/bootstrap-plus/templates/skeleton.vm
@@ -53,7 +53,7 @@
   <script src="${contextRoot}files/fitnesse/bootstrap-plus/js/bootstrap.js" type="text/javascript"></script>
   <script src="${contextRoot}files/fitnesse/bootstrap-plus/js/respond.js" type="text/javascript"></script>
   <script src="${contextRoot}files/fitnesse/bootstrap-plus/js/bootstrap-plus.js" type="text/javascript"></script>
-  <script src="${contextRoot}files/fitnesse/bootstrap-plus/js/bootstrap-plus-team24.js" type="text/javascript"></script>
+  <script src="${contextRoot}files/fitnesse/bootstrap-plus/js/bootstrap-plus-comakership.js" type="text/javascript"></script>
  </head>
  <body#if( $bodyClass ) class="$bodyClass"#end">
